Replaced easytags by taghighlight.
easytags does not seems to be maintenained anymore.
It contained performance issues when using it with vim7.4

I was using easytags only for colors, because i like having a flashy
solarized theme on my screen.

taghighlight seems the perfect match for replacing easytags.
One problem is that it isn't hosted on GitHub :(

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>