Safely source .chruby.sh
When adding .chruby to my bashrc, i did not check whether the file exist
before sourcing it. This results in an error when the file is not
present. That error leads to my .bashrc not being fully parsed, which is
a shame.

This patch checks if .chruby.sh exists before sourcing it in my .bashrc

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>