Update .zshrc with local zshrc
Currently, i was using zsh only on one of my PC's.
The .zshrc i used was not versioned, which is more difficult to share;

I've decided to fix that, so this patch realigns my local .zshrc with
the one on GitHub.
I've removed the local file, and put a ln -s towards the one in this
repository.