Add fugitive commit helper
Since i am used to run git commit from the commandline, i have to change
my workflow in neovim.
I usually do
- git add -up
- git commit -s
However, in a neovim :terminal, this starts vim in neovim.
Having VIMception is something i want to avoid.

This patch introduces a helper script which performs :Gcommit in the
directory of which the script was invoked.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>