Neovim: better git commit helper based on GIT_EDITOR
Currently, i have to run a custom command, gcommit, from shell to start
commiting into neovim. This is not very convenient since my old workflow
was based on using the git commandline.
Moreover, it invoked fugitive's git commit window, which contains a lot
of custom mappings troubling me a lot.

This patch proposes another way to run commits from within neovim's
shell by using standard git and changing the core.editor variable to do
the trick :)

Workflow would be:
- git config core.editor <the python script>
- git add stuff
- git commit -s (as usual)
This opens the commit message in the current nvim split. once the buffer
is hidden, it is deleted from the buffer list and git is saving the
message.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>