nvim: change tnoremap <Esc> to close FZF as expected
FZF popup is shown when trigger keymaps such as <leader>r.
However, in neovim, the window is not closed when entering <Esc>.
This is different from vim behaviour, and disturbing to me.

Since FZF is invoked in term mode and i have rebinded <Esc>, FZF process
do not receive the <Esc> key anymore. This is why it is not closed.

Fix this by sending <Esc> before sending <C-\><C-n> in terminal mode.