spacemacs: layers: add development layer
This patch adds my first customization layer for spacemacs.
There is a lot of boilerplate into the Readme and the packages.el but
it helps me since i'm quite new to the layer system.

This patch also removes dtrt-indent from my additional packages since
it is now into the makohoek-development layer.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>